% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracy&Precision.R
\name{Accuracy & Precision}
\alias{Accuracy & Precision}
\alias{AP}
\title{AP}
\usage{
AP(x, y)
}
\arguments{
\item{x}{The value of the analyzed concentration}

\item{y}{A vector of the dependent variable Y}
}
\value{
Assessment regarding accuracy and precision
}
\description{
The function analyzes the accuracy and precision of the data provided
}
\examples{
Test with TCDF MELDD methodology. Data obtained from (SICUPIRA,2019)
x<- 0.02
y<- c(233153.0,257269.0,307816.0,256882.0,291418.0,264616.0,262520.0)
AP(x,y)
}
\author{
Igor Samuel Mendes
}
