\name{chemometrics-package}
\alias{chemometrics-package}
\alias{chemometrics}
\docType{chemometrics}
\title{
This package is the R companion to the book "Introduction to Multivariate 
Statistical Analysis in Chemometrics" written by K. Varmuza and P. Filzmoser, CRC Press
}
\description{
Included are functions for multivariate statistical methods, tools for diagnostics,
multivariate calibration, cross validation and bootstrap, clustering, etc.
}
\details{
\tabular{ll}{
Package: \tab chemometrics\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2007-11-09\cr
License: \tab GPL version 2 or newer\cr
}
The package can be used to verify the examples in the book. It can also be
used to analyze own data.
}
\author{ P. Filzmoser <P.Filzmoser@tuwien.ac.at} 
\references{
K. Varmuza and P. Filzmoser: Introduction to Multivariate Statistical Analysis
in Chemometrics. CRC Press. To appear.
}
\keyword{ package }
