% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp.R
\name{Comparison}
\alias{Comparison}
\alias{COMP}
\title{COMP}
\usage{
COMP(x)
}
\arguments{
\item{x}{is a data frame}
}
\value{
Test of comparations
}
\description{
The function provides the graph of the fortified sample curve, as well as
that of your white sample, for visual comparison purposes regarding selectivity
of analysis
}
\examples{
# Entrace of datas
DT = c(25, 17, 27, 21, 15,
          10, -2, 12, 4, 16,
         18, 8, 4, 14, 6,
         23, 29, 25, 35,33,
          11, 23, 5, 17, 9,
          8, -6, 6, 0, 2)

# Data processing
(dats = data.frame(trat = factor(rep(c('A','B','C','D','E','Controle'),
                                     each=5)), DT))
attach(dats)
COMP (dats)
}
\author{
Igor Samuel Mendes
}
