\name{glass.grp}
\alias{glass.grp}
\docType{data}
\title{ glass types of the glass data }
\description{
13 different measurements for 180 archaeological glass vessels from
different groups are included. These groups are certain types of
glasses.
}
\usage{data(glass.grp)}
\format{
  The format is:
 num [1:180] 1 1 1 1 1 1 1 1 1 1 ...
}
\details{
This is a vector with 180 elements referring to the groups.
}
\source{
Janssen, K.H.A., De Raedt, I., Schalm, O., Veeckman, J.: Microchim. Acta 15 (suppl.)
(1998) 253-267. Compositions of 15th - 17th century archaeological glass vessels
excavated in Antwerp.
}
\references{
K. Varmuza and P. Filzmoser: Introduction to Multivariate Statistical Analysis
in Chemometrics. CRC Press. To appear. }
}
\examples{
data(glass.grp)
str(glass.grp)
}
\keyword{datasets}
