% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDLQ.R
\name{Limit of detection and quantification}
\alias{Limit of detection and quantification}
\alias{LDLQ}
\title{LDLQ}
\usage{
LDLQ(x, y)
}
\arguments{
\item{x}{Value of the fortified area}

\item{y}{white area value}
}
\value{
Values of LD e LQ
}
\description{
The function informs the analyte's Limit of Detection and its Limit of
Quantification
}
\examples{
x = 49.480
y = 17.054
LDLQ(x,y)

}
\author{
Igor Samuel Mendes
}
