\name{Phenyl}
\alias{Phenyl}
\docType{data}
\title{Phenyl data set}
\description{
The data consist of mass spectra from 600 chemical compounds,
where 300 contain a phenyl substructure (group 1) and 300
compounds do not contain this substructure (group 2). The 
mass spectra have been transformed to 658 variables, containing
the mass spectral features. The 2 groups are coded as -1 (group 1)
and +1 (group 2), and is provided as first last variable. 
}
\usage{data(Phenyl)}
\format{
  A data frame with 600 observations on the following 659 variables.
  \describe{
    \item{\code{grp}}{a numeric vector}
    \item{\code{spec.V1}}{a numeric vector}
    \item{\code{spec.V2}}{a numeric vector}
    \item{\code{spec.V3}}{a numeric vector}
    \item{\code{spec.V4}}{a numeric vector}
    \item{\code{spec.V5}}{a numeric vector}
    \item{\code{spec.V6}}{a numeric vector}
    \item{\code{spec.V7}}{a numeric vector}
    \item{\code{spec.V8}}{a numeric vector}
    \item{\code{spec.V9}}{a numeric vector}
    \item{\code{spec.V10}}{a numeric vector}
    \item{\code{spec.V11}}{a numeric vector}
    \item{\code{spec.V12}}{a numeric vector}
    \item{\code{spec.V13}}{a numeric vector}
    \item{\code{spec.V14}}{a numeric vector}
    \item{\code{spec.V15}}{a numeric vector}
    \item{\code{spec.V16}}{a numeric vector}
    \item{\code{spec.V17}}{a numeric vector}
    \item{\code{spec.V18}}{a numeric vector}
    \item{\code{spec.V19}}{a numeric vector}
    \item{\code{spec.V20}}{a numeric vector}
    \item{\code{spec.V21}}{a numeric vector}
    \item{\code{spec.V22}}{a numeric vector}
    \item{\code{spec.V23}}{a numeric vector}
    \item{\code{spec.V24}}{a numeric vector}
    \item{\code{spec.V25}}{a numeric vector}
    \item{\code{spec.V26}}{a numeric vector}
    \item{\code{spec.V27}}{a numeric vector}
    \item{\code{spec.V28}}{a numeric vector}
    \item{\code{spec.V29}}{a numeric vector}
    \item{\code{spec.V30}}{a numeric vector}
    \item{\code{spec.V31}}{a numeric vector}
    \item{\code{spec.V32}}{a numeric vector}
    \item{\code{spec.V33}}{a numeric vector}
    \item{\code{spec.V34}}{a numeric vector}
    \item{\code{spec.V35}}{a numeric vector}
    \item{\code{spec.V36}}{a numeric vector}
    \item{\code{spec.V37}}{a numeric vector}
    \item{\code{spec.V38}}{a numeric vector}
    \item{\code{spec.V39}}{a numeric vector}
    \item{\code{spec.V40}}{a numeric vector}
    \item{\code{spec.V41}}{a numeric vector}
    \item{\code{spec.V42}}{a numeric vector}
    \item{\code{spec.V43}}{a numeric vector}
    \item{\code{spec.V44}}{a numeric vector}
    \item{\code{spec.V45}}{a numeric vector}
    \item{\code{spec.V46}}{a numeric vector}
    \item{\code{spec.V47}}{a numeric vector}
    \item{\code{spec.V48}}{a numeric vector}
    \item{\code{spec.V49}}{a numeric vector}
    \item{\code{spec.V50}}{a numeric vector}
    \item{\code{spec.V51}}{a numeric vector}
    \item{\code{spec.V52}}{a numeric vector}
    \item{\code{spec.V53}}{a numeric vector}
    \item{\code{spec.V54}}{a numeric vector}
    \item{\code{spec.V55}}{a numeric vector}
    \item{\code{spec.V56}}{a numeric vector}
    \item{\code{spec.V57}}{a numeric vector}
    \item{\code{spec.V58}}{a numeric vector}
    \item{\code{spec.V59}}{a numeric vector}
    \item{\code{spec.V60}}{a numeric vector}
    \item{\code{spec.V61}}{a numeric vector}
    \item{\code{spec.V62}}{a numeric vector}
    \item{\code{spec.V63}}{a numeric vector}
    \item{\code{spec.V64}}{a numeric vector}
    \item{\code{spec.V65}}{a numeric vector}
    \item{\code{spec.V66}}{a numeric vector}
    \item{\code{spec.V67}}{a numeric vector}
    \item{\code{spec.V68}}{a numeric vector}
    \item{\code{spec.V69}}{a numeric vector}
    \item{\code{spec.V70}}{a numeric vector}
    \item{\code{spec.V71}}{a numeric vector}
    \item{\code{spec.V72}}{a numeric vector}
    \item{\code{spec.V73}}{a numeric vector}
    \item{\code{spec.V74}}{a numeric vector}
    \item{\code{spec.V75}}{a numeric vector}
    \item{\code{spec.V76}}{a numeric vector}
    \item{\code{spec.V77}}{a numeric vector}
    \item{\code{spec.V78}}{a numeric vector}
    \item{\code{spec.V79}}{a numeric vector}
    \item{\code{spec.V80}}{a numeric vector}
    \item{\code{spec.V81}}{a numeric vector}
    \item{\code{spec.V82}}{a numeric vector}
    \item{\code{spec.V83}}{a numeric vector}
    \item{\code{spec.V84}}{a numeric vector}
    \item{\code{spec.V85}}{a numeric vector}
    \item{\code{spec.V86}}{a numeric vector}
    \item{\code{spec.V87}}{a numeric vector}
    \item{\code{spec.V88}}{a numeric vector}
    \item{\code{spec.V89}}{a numeric vector}
    \item{\code{spec.V90}}{a numeric vector}
    \item{\code{spec.V91}}{a numeric vector}
    \item{\code{spec.V92}}{a numeric vector}
    \item{\code{spec.V93}}{a numeric vector}
    \item{\code{spec.V94}}{a numeric vector}
    \item{\code{spec.V95}}{a numeric vector}
    \item{\code{spec.V96}}{a numeric vector}
    \item{\code{spec.V97}}{a numeric vector}
    \item{\code{spec.V98}}{a numeric vector}
    \item{\code{spec.V99}}{a numeric vector}
    \item{\code{spec.V100}}{a numeric vector}
    \item{\code{spec.V101}}{a numeric vector}
    \item{\code{spec.V102}}{a numeric vector}
    \item{\code{spec.V103}}{a numeric vector}
    \item{\code{spec.V104}}{a numeric vector}
    \item{\code{spec.V105}}{a numeric vector}
    \item{\code{spec.V106}}{a numeric vector}
    \item{\code{spec.V107}}{a numeric vector}
    \item{\code{spec.V108}}{a numeric vector}
    \item{\code{spec.V109}}{a numeric vector}
    \item{\code{spec.V110}}{a numeric vector}
    \item{\code{spec.V111}}{a numeric vector}
    \item{\code{spec.V112}}{a numeric vector}
    \item{\code{spec.V113}}{a numeric vector}
    \item{\code{spec.V114}}{a numeric vector}
    \item{\code{spec.V115}}{a numeric vector}
    \item{\code{spec.V116}}{a numeric vector}
    \item{\code{spec.V117}}{a numeric vector}
    \item{\code{spec.V118}}{a numeric vector}
    \item{\code{spec.V119}}{a numeric vector}
    \item{\code{spec.V120}}{a numeric vector}
    \item{\code{spec.V121}}{a numeric vector}
    \item{\code{spec.V122}}{a numeric vector}
    \item{\code{spec.V123}}{a numeric vector}
    \item{\code{spec.V124}}{a numeric vector}
    \item{\code{spec.V125}}{a numeric vector}
    \item{\code{spec.V126}}{a numeric vector}
    \item{\code{spec.V127}}{a numeric vector}
    \item{\code{spec.V128}}{a numeric vector}
    \item{\code{spec.V129}}{a numeric vector}
    \item{\code{spec.V130}}{a numeric vector}
    \item{\code{spec.V131}}{a numeric vector}
    \item{\code{spec.V132}}{a numeric vector}
    \item{\code{spec.V133}}{a numeric vector}
    \item{\code{spec.V134}}{a numeric vector}
    \item{\code{spec.V135}}{a numeric vector}
    \item{\code{spec.V136}}{a numeric vector}
    \item{\code{spec.V137}}{a numeric vector}
    \item{\code{spec.V138}}{a numeric vector}
    \item{\code{spec.V139}}{a numeric vector}
    \item{\code{spec.V140}}{a numeric vector}
    \item{\code{spec.V141}}{a numeric vector}
    \item{\code{spec.V142}}{a numeric vector}
    \item{\code{spec.V143}}{a numeric vector}
    \item{\code{spec.V144}}{a numeric vector}
    \item{\code{spec.V145}}{a numeric vector}
    \item{\code{spec.V146}}{a numeric vector}
    \item{\code{spec.V147}}{a numeric vector}
    \item{\code{spec.V148}}{a numeric vector}
    \item{\code{spec.V149}}{a numeric vector}
    \item{\code{spec.V150}}{a numeric vector}
    \item{\code{spec.V151}}{a numeric vector}
    \item{\code{spec.V152}}{a numeric vector}
    \item{\code{spec.V153}}{a numeric vector}
    \item{\code{spec.V154}}{a numeric vector}
    \item{\code{spec.V155}}{a numeric vector}
    \item{\code{spec.V156}}{a numeric vector}
    \item{\code{spec.V157}}{a numeric vector}
    \item{\code{spec.V158}}{a numeric vector}
    \item{\code{spec.V159}}{a numeric vector}
    \item{\code{spec.V160}}{a numeric vector}
    \item{\code{spec.V161}}{a numeric vector}
    \item{\code{spec.V162}}{a numeric vector}
    \item{\code{spec.V163}}{a numeric vector}
    \item{\code{spec.V164}}{a numeric vector}
    \item{\code{spec.V165}}{a numeric vector}
    \item{\code{spec.V166}}{a numeric vector}
    \item{\code{spec.V167}}{a numeric vector}
    \item{\code{spec.V168}}{a numeric vector}
    \item{\code{spec.V169}}{a numeric vector}
    \item{\code{spec.V170}}{a numeric vector}
    \item{\code{spec.V171}}{a numeric vector}
    \item{\code{spec.V172}}{a numeric vector}
    \item{\code{spec.V173}}{a numeric vector}
    \item{\code{spec.V174}}{a numeric vector}
    \item{\code{spec.V175}}{a numeric vector}
    \item{\code{spec.V176}}{a numeric vector}
    \item{\code{spec.V177}}{a numeric vector}
    \item{\code{spec.V178}}{a numeric vector}
    \item{\code{spec.V179}}{a numeric vector}
    \item{\code{spec.V180}}{a numeric vector}
    \item{\code{spec.V181}}{a numeric vector}
    \item{\code{spec.V182}}{a numeric vector}
    \item{\code{spec.V183}}{a numeric vector}
    \item{\code{spec.V184}}{a numeric vector}
    \item{\code{spec.V185}}{a numeric vector}
    \item{\code{spec.V186}}{a numeric vector}
    \item{\code{spec.V187}}{a numeric vector}
    \item{\code{spec.V188}}{a numeric vector}
    \item{\code{spec.V189}}{a numeric vector}
    \item{\code{spec.V190}}{a numeric vector}
    \item{\code{spec.V191}}{a numeric vector}
    \item{\code{spec.V192}}{a numeric vector}
    \item{\code{spec.V193}}{a numeric vector}
    \item{\code{spec.V194}}{a numeric vector}
    \item{\code{spec.V195}}{a numeric vector}
    \item{\code{spec.V196}}{a numeric vector}
    \item{\code{spec.V197}}{a numeric vector}
    \item{\code{spec.V198}}{a numeric vector}
    \item{\code{spec.V199}}{a numeric vector}
    \item{\code{spec.V200}}{a numeric vector}
    \item{\code{spec.V201}}{a numeric vector}
    \item{\code{spec.V202}}{a numeric vector}
    \item{\code{spec.V203}}{a numeric vector}
    \item{\code{spec.V204}}{a numeric vector}
    \item{\code{spec.V205}}{a numeric vector}
    \item{\code{spec.V206}}{a numeric vector}
    \item{\code{spec.V207}}{a numeric vector}
    \item{\code{spec.V208}}{a numeric vector}
    \item{\code{spec.V209}}{a numeric vector}
    \item{\code{spec.V210}}{a numeric vector}
    \item{\code{spec.V211}}{a numeric vector}
    \item{\code{spec.V212}}{a numeric vector}
    \item{\code{spec.V213}}{a numeric vector}
    \item{\code{spec.V214}}{a numeric vector}
    \item{\code{spec.V215}}{a numeric vector}
    \item{\code{spec.V216}}{a numeric vector}
    \item{\code{spec.V217}}{a numeric vector}
    \item{\code{spec.V218}}{a numeric vector}
    \item{\code{spec.V219}}{a numeric vector}
    \item{\code{spec.V220}}{a numeric vector}
    \item{\code{spec.V221}}{a numeric vector}
    \item{\code{spec.V222}}{a numeric vector}
    \item{\code{spec.V223}}{a numeric vector}
    \item{\code{spec.V224}}{a numeric vector}
    \item{\code{spec.V225}}{a numeric vector}
    \item{\code{spec.V226}}{a numeric vector}
    \item{\code{spec.V227}}{a numeric vector}
    \item{\code{spec.V228}}{a numeric vector}
    \item{\code{spec.V229}}{a numeric vector}
    \item{\code{spec.V230}}{a numeric vector}
    \item{\code{spec.V231}}{a numeric vector}
    \item{\code{spec.V232}}{a numeric vector}
    \item{\code{spec.V233}}{a numeric vector}
    \item{\code{spec.V234}}{a numeric vector}
    \item{\code{spec.V235}}{a numeric vector}
    \item{\code{spec.V236}}{a numeric vector}
    \item{\code{spec.V237}}{a numeric vector}
    \item{\code{spec.V238}}{a numeric vector}
    \item{\code{spec.V239}}{a numeric vector}
    \item{\code{spec.V240}}{a numeric vector}
    \item{\code{spec.V241}}{a numeric vector}
    \item{\code{spec.V242}}{a numeric vector}
    \item{\code{spec.V243}}{a numeric vector}
    \item{\code{spec.V244}}{a numeric vector}
    \item{\code{spec.V245}}{a numeric vector}
    \item{\code{spec.V246}}{a numeric vector}
    \item{\code{spec.V247}}{a numeric vector}
    \item{\code{spec.V248}}{a numeric vector}
    \item{\code{spec.V249}}{a numeric vector}
    \item{\code{spec.V250}}{a numeric vector}
    \item{\code{spec.V251}}{a numeric vector}
    \item{\code{spec.V252}}{a numeric vector}
    \item{\code{spec.V253}}{a numeric vector}
    \item{\code{spec.V254}}{a numeric vector}
    \item{\code{spec.V255}}{a numeric vector}
    \item{\code{spec.V256}}{a numeric vector}
    \item{\code{spec.V257}}{a numeric vector}
    \item{\code{spec.V258}}{a numeric vector}
    \item{\code{spec.V259}}{a numeric vector}
    \item{\code{spec.V260}}{a numeric vector}
    \item{\code{spec.V261}}{a numeric vector}
    \item{\code{spec.V262}}{a numeric vector}
    \item{\code{spec.V263}}{a numeric vector}
    \item{\code{spec.V264}}{a numeric vector}
    \item{\code{spec.V265}}{a numeric vector}
    \item{\code{spec.V266}}{a numeric vector}
    \item{\code{spec.V267}}{a numeric vector}
    \item{\code{spec.V268}}{a numeric vector}
    \item{\code{spec.V269}}{a numeric vector}
    \item{\code{spec.V270}}{a numeric vector}
    \item{\code{spec.V271}}{a numeric vector}
    \item{\code{spec.V272}}{a numeric vector}
    \item{\code{spec.V273}}{a numeric vector}
    \item{\code{spec.V274}}{a numeric vector}
    \item{\code{spec.V275}}{a numeric vector}
    \item{\code{spec.V276}}{a numeric vector}
    \item{\code{spec.V277}}{a numeric vector}
    \item{\code{spec.V278}}{a numeric vector}
    \item{\code{spec.V279}}{a numeric vector}
    \item{\code{spec.V280}}{a numeric vector}
    \item{\code{spec.V281}}{a numeric vector}
    \item{\code{spec.V282}}{a numeric vector}
    \item{\code{spec.V283}}{a numeric vector}
    \item{\code{spec.V284}}{a numeric vector}
    \item{\code{spec.V285}}{a numeric vector}
    \item{\code{spec.V286}}{a numeric vector}
    \item{\code{spec.V287}}{a numeric vector}
    \item{\code{spec.V288}}{a numeric vector}
    \item{\code{spec.V289}}{a numeric vector}
    \item{\code{spec.V290}}{a numeric vector}
    \item{\code{spec.V291}}{a numeric vector}
    \item{\code{spec.V292}}{a numeric vector}
    \item{\code{spec.V293}}{a numeric vector}
    \item{\code{spec.V294}}{a numeric vector}
    \item{\code{spec.V295}}{a numeric vector}
    \item{\code{spec.V296}}{a numeric vector}
    \item{\code{spec.V297}}{a numeric vector}
    \item{\code{spec.V298}}{a numeric vector}
    \item{\code{spec.V299}}{a numeric vector}
    \item{\code{spec.V300}}{a numeric vector}
    \item{\code{spec.V301}}{a numeric vector}
    \item{\code{spec.V302}}{a numeric vector}
    \item{\code{spec.V303}}{a numeric vector}
    \item{\code{spec.V304}}{a numeric vector}
    \item{\code{spec.V305}}{a numeric vector}
    \item{\code{spec.V306}}{a numeric vector}
    \item{\code{spec.V307}}{a numeric vector}
    \item{\code{spec.V308}}{a numeric vector}
    \item{\code{spec.V309}}{a numeric vector}
    \item{\code{spec.V310}}{a numeric vector}
    \item{\code{spec.V311}}{a numeric vector}
    \item{\code{spec.V312}}{a numeric vector}
    \item{\code{spec.V313}}{a numeric vector}
    \item{\code{spec.V314}}{a numeric vector}
    \item{\code{spec.V315}}{a numeric vector}
    \item{\code{spec.V316}}{a numeric vector}
    \item{\code{spec.V317}}{a numeric vector}
    \item{\code{spec.V318}}{a numeric vector}
    \item{\code{spec.V319}}{a numeric vector}
    \item{\code{spec.V320}}{a numeric vector}
    \item{\code{spec.V321}}{a numeric vector}
    \item{\code{spec.V322}}{a numeric vector}
    \item{\code{spec.V323}}{a numeric vector}
    \item{\code{spec.V324}}{a numeric vector}
    \item{\code{spec.V325}}{a numeric vector}
    \item{\code{spec.V326}}{a numeric vector}
    \item{\code{spec.V327}}{a numeric vector}
    \item{\code{spec.V328}}{a numeric vector}
    \item{\code{spec.V329}}{a numeric vector}
    \item{\code{spec.V330}}{a numeric vector}
    \item{\code{spec.V331}}{a numeric vector}
    \item{\code{spec.V332}}{a numeric vector}
    \item{\code{spec.V333}}{a numeric vector}
    \item{\code{spec.V334}}{a numeric vector}
    \item{\code{spec.V335}}{a numeric vector}
    \item{\code{spec.V336}}{a numeric vector}
    \item{\code{spec.V337}}{a numeric vector}
    \item{\code{spec.V338}}{a numeric vector}
    \item{\code{spec.V339}}{a numeric vector}
    \item{\code{spec.V340}}{a numeric vector}
    \item{\code{spec.V341}}{a numeric vector}
    \item{\code{spec.V342}}{a numeric vector}
    \item{\code{spec.V343}}{a numeric vector}
    \item{\code{spec.V344}}{a numeric vector}
    \item{\code{spec.V345}}{a numeric vector}
    \item{\code{spec.V346}}{a numeric vector}
    \item{\code{spec.V347}}{a numeric vector}
    \item{\code{spec.V348}}{a numeric vector}
    \item{\code{spec.V349}}{a numeric vector}
    \item{\code{spec.V350}}{a numeric vector}
    \item{\code{spec.V351}}{a numeric vector}
    \item{\code{spec.V352}}{a numeric vector}
    \item{\code{spec.V353}}{a numeric vector}
    \item{\code{spec.V354}}{a numeric vector}
    \item{\code{spec.V355}}{a numeric vector}
    \item{\code{spec.V356}}{a numeric vector}
    \item{\code{spec.V357}}{a numeric vector}
    \item{\code{spec.V358}}{a numeric vector}
    \item{\code{spec.V359}}{a numeric vector}
    \item{\code{spec.V360}}{a numeric vector}
    \item{\code{spec.V361}}{a numeric vector}
    \item{\code{spec.V362}}{a numeric vector}
    \item{\code{spec.V363}}{a numeric vector}
    \item{\code{spec.V364}}{a numeric vector}
    \item{\code{spec.V365}}{a numeric vector}
    \item{\code{spec.V366}}{a numeric vector}
    \item{\code{spec.V367}}{a numeric vector}
    \item{\code{spec.V368}}{a numeric vector}
    \item{\code{spec.V369}}{a numeric vector}
    \item{\code{spec.V370}}{a numeric vector}
    \item{\code{spec.V371}}{a numeric vector}
    \item{\code{spec.V372}}{a numeric vector}
    \item{\code{spec.V373}}{a numeric vector}
    \item{\code{spec.V374}}{a numeric vector}
    \item{\code{spec.V375}}{a numeric vector}
    \item{\code{spec.V376}}{a numeric vector}
    \item{\code{spec.V377}}{a numeric vector}
    \item{\code{spec.V378}}{a numeric vector}
    \item{\code{spec.V379}}{a numeric vector}
    \item{\code{spec.V380}}{a numeric vector}
    \item{\code{spec.V381}}{a numeric vector}
    \item{\code{spec.V382}}{a numeric vector}
    \item{\code{spec.V383}}{a numeric vector}
    \item{\code{spec.V384}}{a numeric vector}
    \item{\code{spec.V385}}{a numeric vector}
    \item{\code{spec.V386}}{a numeric vector}
    \item{\code{spec.V387}}{a numeric vector}
    \item{\code{spec.V388}}{a numeric vector}
    \item{\code{spec.V389}}{a numeric vector}
    \item{\code{spec.V390}}{a numeric vector}
    \item{\code{spec.V391}}{a numeric vector}
    \item{\code{spec.V392}}{a numeric vector}
    \item{\code{spec.V393}}{a numeric vector}
    \item{\code{spec.V394}}{a numeric vector}
    \item{\code{spec.V395}}{a numeric vector}
    \item{\code{spec.V396}}{a numeric vector}
    \item{\code{spec.V397}}{a numeric vector}
    \item{\code{spec.V398}}{a numeric vector}
    \item{\code{spec.V399}}{a numeric vector}
    \item{\code{spec.V400}}{a numeric vector}
    \item{\code{spec.V401}}{a numeric vector}
    \item{\code{spec.V402}}{a numeric vector}
    \item{\code{spec.V403}}{a numeric vector}
    \item{\code{spec.V404}}{a numeric vector}
    \item{\code{spec.V405}}{a numeric vector}
    \item{\code{spec.V406}}{a numeric vector}
    \item{\code{spec.V407}}{a numeric vector}
    \item{\code{spec.V408}}{a numeric vector}
    \item{\code{spec.V409}}{a numeric vector}
    \item{\code{spec.V410}}{a numeric vector}
    \item{\code{spec.V411}}{a numeric vector}
    \item{\code{spec.V412}}{a numeric vector}
    \item{\code{spec.V413}}{a numeric vector}
    \item{\code{spec.V414}}{a numeric vector}
    \item{\code{spec.V415}}{a numeric vector}
    \item{\code{spec.V416}}{a numeric vector}
    \item{\code{spec.V417}}{a numeric vector}
    \item{\code{spec.V418}}{a numeric vector}
    \item{\code{spec.V419}}{a numeric vector}
    \item{\code{spec.V420}}{a numeric vector}
    \item{\code{spec.V421}}{a numeric vector}
    \item{\code{spec.V422}}{a numeric vector}
    \item{\code{spec.V423}}{a numeric vector}
    \item{\code{spec.V424}}{a numeric vector}
    \item{\code{spec.V425}}{a numeric vector}
    \item{\code{spec.V426}}{a numeric vector}
    \item{\code{spec.V427}}{a numeric vector}
    \item{\code{spec.V428}}{a numeric vector}
    \item{\code{spec.V429}}{a numeric vector}
    \item{\code{spec.V430}}{a numeric vector}
    \item{\code{spec.V431}}{a numeric vector}
    \item{\code{spec.V432}}{a numeric vector}
    \item{\code{spec.V433}}{a numeric vector}
    \item{\code{spec.V434}}{a numeric vector}
    \item{\code{spec.V435}}{a numeric vector}
    \item{\code{spec.V436}}{a numeric vector}
    \item{\code{spec.V437}}{a numeric vector}
    \item{\code{spec.V438}}{a numeric vector}
    \item{\code{spec.V439}}{a numeric vector}
    \item{\code{spec.V440}}{a numeric vector}
    \item{\code{spec.V441}}{a numeric vector}
    \item{\code{spec.V442}}{a numeric vector}
    \item{\code{spec.V443}}{a numeric vector}
    \item{\code{spec.V444}}{a numeric vector}
    \item{\code{spec.V445}}{a numeric vector}
    \item{\code{spec.V446}}{a numeric vector}
    \item{\code{spec.V447}}{a numeric vector}
    \item{\code{spec.V448}}{a numeric vector}
    \item{\code{spec.V449}}{a numeric vector}
    \item{\code{spec.V450}}{a numeric vector}
    \item{\code{spec.V451}}{a numeric vector}
    \item{\code{spec.V452}}{a numeric vector}
    \item{\code{spec.V453}}{a numeric vector}
    \item{\code{spec.V454}}{a numeric vector}
    \item{\code{spec.V455}}{a numeric vector}
    \item{\code{spec.V456}}{a numeric vector}
    \item{\code{spec.V457}}{a numeric vector}
    \item{\code{spec.V458}}{a numeric vector}
    \item{\code{spec.V459}}{a numeric vector}
    \item{\code{spec.V460}}{a numeric vector}
    \item{\code{spec.V461}}{a numeric vector}
    \item{\code{spec.V462}}{a numeric vector}
    \item{\code{spec.V463}}{a numeric vector}
    \item{\code{spec.V464}}{a numeric vector}
    \item{\code{spec.V465}}{a numeric vector}
    \item{\code{spec.V466}}{a numeric vector}
    \item{\code{spec.V467}}{a numeric vector}
    \item{\code{spec.V468}}{a numeric vector}
    \item{\code{spec.V469}}{a numeric vector}
    \item{\code{spec.V470}}{a numeric vector}
    \item{\code{spec.V471}}{a numeric vector}
    \item{\code{spec.V472}}{a numeric vector}
    \item{\code{spec.V473}}{a numeric vector}
    \item{\code{spec.V474}}{a numeric vector}
    \item{\code{spec.V475}}{a numeric vector}
    \item{\code{spec.V476}}{a numeric vector}
    \item{\code{spec.V477}}{a numeric vector}
    \item{\code{spec.V478}}{a numeric vector}
    \item{\code{spec.V479}}{a numeric vector}
    \item{\code{spec.V480}}{a numeric vector}
    \item{\code{spec.V481}}{a numeric vector}
    \item{\code{spec.V482}}{a numeric vector}
    \item{\code{spec.V483}}{a numeric vector}
    \item{\code{spec.V484}}{a numeric vector}
    \item{\code{spec.V485}}{a numeric vector}
    \item{\code{spec.V486}}{a numeric vector}
    \item{\code{spec.V487}}{a numeric vector}
    \item{\code{spec.V488}}{a numeric vector}
    \item{\code{spec.V489}}{a numeric vector}
    \item{\code{spec.V490}}{a numeric vector}
    \item{\code{spec.V491}}{a numeric vector}
    \item{\code{spec.V492}}{a numeric vector}
    \item{\code{spec.V493}}{a numeric vector}
    \item{\code{spec.V494}}{a numeric vector}
    \item{\code{spec.V495}}{a numeric vector}
    \item{\code{spec.V496}}{a numeric vector}
    \item{\code{spec.V497}}{a numeric vector}
    \item{\code{spec.V498}}{a numeric vector}
    \item{\code{spec.V499}}{a numeric vector}
    \item{\code{spec.V500}}{a numeric vector}
    \item{\code{spec.V501}}{a numeric vector}
    \item{\code{spec.V502}}{a numeric vector}
    \item{\code{spec.V503}}{a numeric vector}
    \item{\code{spec.V504}}{a numeric vector}
    \item{\code{spec.V505}}{a numeric vector}
    \item{\code{spec.V506}}{a numeric vector}
    \item{\code{spec.V507}}{a numeric vector}
    \item{\code{spec.V508}}{a numeric vector}
    \item{\code{spec.V509}}{a numeric vector}
    \item{\code{spec.V510}}{a numeric vector}
    \item{\code{spec.V511}}{a numeric vector}
    \item{\code{spec.V512}}{a numeric vector}
    \item{\code{spec.V513}}{a numeric vector}
    \item{\code{spec.V514}}{a numeric vector}
    \item{\code{spec.V515}}{a numeric vector}
    \item{\code{spec.V516}}{a numeric vector}
    \item{\code{spec.V517}}{a numeric vector}
    \item{\code{spec.V518}}{a numeric vector}
    \item{\code{spec.V519}}{a numeric vector}
    \item{\code{spec.V520}}{a numeric vector}
    \item{\code{spec.V521}}{a numeric vector}
    \item{\code{spec.V522}}{a numeric vector}
    \item{\code{spec.V523}}{a numeric vector}
    \item{\code{spec.V524}}{a numeric vector}
    \item{\code{spec.V525}}{a numeric vector}
    \item{\code{spec.V526}}{a numeric vector}
    \item{\code{spec.V527}}{a numeric vector}
    \item{\code{spec.V528}}{a numeric vector}
    \item{\code{spec.V529}}{a numeric vector}
    \item{\code{spec.V530}}{a numeric vector}
    \item{\code{spec.V531}}{a numeric vector}
    \item{\code{spec.V532}}{a numeric vector}
    \item{\code{spec.V533}}{a numeric vector}
    \item{\code{spec.V534}}{a numeric vector}
    \item{\code{spec.V535}}{a numeric vector}
    \item{\code{spec.V536}}{a numeric vector}
    \item{\code{spec.V537}}{a numeric vector}
    \item{\code{spec.V538}}{a numeric vector}
    \item{\code{spec.V539}}{a numeric vector}
    \item{\code{spec.V540}}{a numeric vector}
    \item{\code{spec.V541}}{a numeric vector}
    \item{\code{spec.V542}}{a numeric vector}
    \item{\code{spec.V543}}{a numeric vector}
    \item{\code{spec.V544}}{a numeric vector}
    \item{\code{spec.V545}}{a numeric vector}
    \item{\code{spec.V546}}{a numeric vector}
    \item{\code{spec.V547}}{a numeric vector}
    \item{\code{spec.V548}}{a numeric vector}
    \item{\code{spec.V549}}{a numeric vector}
    \item{\code{spec.V550}}{a numeric vector}
    \item{\code{spec.V551}}{a numeric vector}
    \item{\code{spec.V552}}{a numeric vector}
    \item{\code{spec.V553}}{a numeric vector}
    \item{\code{spec.V554}}{a numeric vector}
    \item{\code{spec.V555}}{a numeric vector}
    \item{\code{spec.V556}}{a numeric vector}
    \item{\code{spec.V557}}{a numeric vector}
    \item{\code{spec.V558}}{a numeric vector}
    \item{\code{spec.V559}}{a numeric vector}
    \item{\code{spec.V560}}{a numeric vector}
    \item{\code{spec.V561}}{a numeric vector}
    \item{\code{spec.V562}}{a numeric vector}
    \item{\code{spec.V563}}{a numeric vector}
    \item{\code{spec.V564}}{a numeric vector}
    \item{\code{spec.V565}}{a numeric vector}
    \item{\code{spec.V566}}{a numeric vector}
    \item{\code{spec.V567}}{a numeric vector}
    \item{\code{spec.V568}}{a numeric vector}
    \item{\code{spec.V569}}{a numeric vector}
    \item{\code{spec.V570}}{a numeric vector}
    \item{\code{spec.V571}}{a numeric vector}
    \item{\code{spec.V572}}{a numeric vector}
    \item{\code{spec.V573}}{a numeric vector}
    \item{\code{spec.V574}}{a numeric vector}
    \item{\code{spec.V575}}{a numeric vector}
    \item{\code{spec.V576}}{a numeric vector}
    \item{\code{spec.V577}}{a numeric vector}
    \item{\code{spec.V578}}{a numeric vector}
    \item{\code{spec.V579}}{a numeric vector}
    \item{\code{spec.V580}}{a numeric vector}
    \item{\code{spec.V581}}{a numeric vector}
    \item{\code{spec.V582}}{a numeric vector}
    \item{\code{spec.V583}}{a numeric vector}
    \item{\code{spec.V584}}{a numeric vector}
    \item{\code{spec.V585}}{a numeric vector}
    \item{\code{spec.V586}}{a numeric vector}
    \item{\code{spec.V587}}{a numeric vector}
    \item{\code{spec.V588}}{a numeric vector}
    \item{\code{spec.V589}}{a numeric vector}
    \item{\code{spec.V590}}{a numeric vector}
    \item{\code{spec.V591}}{a numeric vector}
    \item{\code{spec.V592}}{a numeric vector}
    \item{\code{spec.V593}}{a numeric vector}
    \item{\code{spec.V594}}{a numeric vector}
    \item{\code{spec.V595}}{a numeric vector}
    \item{\code{spec.V596}}{a numeric vector}
    \item{\code{spec.V597}}{a numeric vector}
    \item{\code{spec.V598}}{a numeric vector}
    \item{\code{spec.V599}}{a numeric vector}
    \item{\code{spec.V600}}{a numeric vector}
    \item{\code{spec.V601}}{a numeric vector}
    \item{\code{spec.V602}}{a numeric vector}
    \item{\code{spec.V603}}{a numeric vector}
    \item{\code{spec.V604}}{a numeric vector}
    \item{\code{spec.V605}}{a numeric vector}
    \item{\code{spec.V606}}{a numeric vector}
    \item{\code{spec.V607}}{a numeric vector}
    \item{\code{spec.V608}}{a numeric vector}
    \item{\code{spec.V609}}{a numeric vector}
    \item{\code{spec.V610}}{a numeric vector}
    \item{\code{spec.V611}}{a numeric vector}
    \item{\code{spec.V612}}{a numeric vector}
    \item{\code{spec.V613}}{a numeric vector}
    \item{\code{spec.V614}}{a numeric vector}
    \item{\code{spec.V615}}{a numeric vector}
    \item{\code{spec.V616}}{a numeric vector}
    \item{\code{spec.V617}}{a numeric vector}
    \item{\code{spec.V618}}{a numeric vector}
    \item{\code{spec.V619}}{a numeric vector}
    \item{\code{spec.V620}}{a numeric vector}
    \item{\code{spec.V621}}{a numeric vector}
    \item{\code{spec.V622}}{a numeric vector}
    \item{\code{spec.V623}}{a numeric vector}
    \item{\code{spec.V624}}{a numeric vector}
    \item{\code{spec.V625}}{a numeric vector}
    \item{\code{spec.V626}}{a numeric vector}
    \item{\code{spec.V627}}{a numeric vector}
    \item{\code{spec.V628}}{a numeric vector}
    \item{\code{spec.V629}}{a numeric vector}
    \item{\code{spec.V630}}{a numeric vector}
    \item{\code{spec.V631}}{a numeric vector}
    \item{\code{spec.V632}}{a numeric vector}
    \item{\code{spec.V633}}{a numeric vector}
    \item{\code{spec.V634}}{a numeric vector}
    \item{\code{spec.V635}}{a numeric vector}
    \item{\code{spec.V636}}{a numeric vector}
    \item{\code{spec.V637}}{a numeric vector}
    \item{\code{spec.V638}}{a numeric vector}
    \item{\code{spec.V639}}{a numeric vector}
    \item{\code{spec.V640}}{a numeric vector}
    \item{\code{spec.V641}}{a numeric vector}
    \item{\code{spec.V642}}{a numeric vector}
    \item{\code{spec.V643}}{a numeric vector}
    \item{\code{spec.V644}}{a numeric vector}
    \item{\code{spec.V645}}{a numeric vector}
    \item{\code{spec.V646}}{a numeric vector}
    \item{\code{spec.V647}}{a numeric vector}
    \item{\code{spec.V648}}{a numeric vector}
    \item{\code{spec.V649}}{a numeric vector}
    \item{\code{spec.V650}}{a numeric vector}
    \item{\code{spec.V651}}{a numeric vector}
    \item{\code{spec.V652}}{a numeric vector}
    \item{\code{spec.V653}}{a numeric vector}
    \item{\code{spec.V654}}{a numeric vector}
    \item{\code{spec.V655}}{a numeric vector}
    \item{\code{spec.V656}}{a numeric vector}
    \item{\code{spec.V657}}{a numeric vector}
    \item{\code{spec.V658}}{a numeric vector}
  }
}
\details{
The data set can be used for classification in high dimensions.
}
\source{
K. Varmuza and P. Filzmoser: Introduction to Multivariate Statistical Analysis
in Chemeometrics. CRC Press, Boca Raton, FL, 2009. }
\references{
K. Varmuza and P. Filzmoser: Introduction to Multivariate Statistical Analysis
in Chemeometrics. CRC Press, Boca Raton, FL, 2009. }
\examples{
data(Phenyl)
str(Phenyl)
}
\keyword{datasets}
