% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivity.R
\name{Selectivity}
\alias{Selectivity}
\alias{SEL}
\title{Sel}
\usage{
SEL(tdp, t, ym, y, z)
}
\arguments{
\item{tdp}{Total data points}

\item{t}{Time,in minutes, of the Chromatographic run}

\item{ym}{Y axys multiplier}

\item{y}{spiked analyte response}

\item{z}{blank sample response}
}
\value{
Graph for selectivity analysis
}
\description{
The function provides the graph of the fortified sample curve, as well as
that of your white sample, for visual comparison purposes regarding selectivity
of analysis
}
\author{
Igor Samuel Mendes
}
