\name{chemometrics-package}
\alias{chemometrics-package}
\alias{chemometrics}
\docType{package}
\title{
This package is the R companion to the book "Introduction to Multivariate 
Statistical Analysis in Chemometrics" written by K. Varmuza and P. Filzmoser (2009).
}
\description{
Included are functions for multivariate statistical methods, tools for diagnostics,
multivariate calibration, cross validation and bootstrap, clustering, etc.
}
\details{
The package can be used to verify the examples in the book. It can also be
used to analyze own data.
}
\author{ P. Filzmoser <P.Filzmoser@tuwien.ac.at} 
\references{
K. Varmuza and P. Filzmoser: Introduction to Multivariate Statistical Analysis
in Chemometrics. CRC Press, Boca Raton, FL, 2009.
}
\keyword{ package }
